Add the cplocn2atm namelist option to turn on/off ocean feedback to FV3atm (#7)
Currently, when running coupled configurations/applications (e.g., HAFS or S2S). There is no available option to control one-way coupling to turn off the ocean model component feedback (e.g., SST) to FV3ATM.

Similar to the cplwav2atm namelist option, add a cplocn2atm namelist option to turn on/off the ocean model component feedback (e.g., SST) to the atmosphere model component. This is useful to control the one-way or two-way atmosphere-ocean coupling, as well as to configure different combinations for the atmosphere-ocean-wave coupling in HAFS and/or other applications.