add semi-lagrangian sedimentation of graupel and update the namelist … (#426)
The semi-Lagrangian sedimentation of graupel is added. The namelist variables controlling the semi-Lagrangian sedimentation of precipitation (rain and graupel) are modified. Now SEDI_SEMI and DECFL is the switch. DECFL is the deformed CFL factor.

Semi-Lagrangian sedimentation of precipitation is added as an option and does not change results before it is turned on.